[spv-in] Split `spv/mod.rs` into two files.
`Frontend::next_block()` is the bulk of the logic, and moving it and
nothing else into a separate file produces an even split in half.

This is, of course, inferior to figuring out some way to make the huge
function smaller, but it should be a cheap benefit to editability
meanwhile.
